17/04/23

+++ views/categoria/editar.php
@@ -1,6 +1,8 @@
 <?php
 include('../../includes/header.php');
 
+require_once('../../models/Categoria.php');
+
 // establecer la conexión a la base de datos
 $conexion = new mysqli('localhost', 'root', '', 'booking a book');
 
@@ -15,7 +17,7 @@ if (isset($_POST['actualizar']) || isset($_POST['eliminar'])) {
   $id = $_POST['id'];
 
   // determinar si se está eliminando la categoría
-  $eliminar = isset($_POST['eliminar']) && $_POST['eliminar'] === 'true';
+$eliminar = isset($_POST['eliminar']);
 
   // si se está actualizando la categoría, obtener el nuevo nombre
   $nombre = null;
@@ -23,28 +25,31 @@ if (isset($_POST['actualizar']) || isset($_POST['eliminar'])) {
     $nombre = $_POST['nombre'];
   }
 
-  try {
-    // actualizar o eliminar la categoría en la base de datos
-    if ($eliminar) {
-      $query = "DELETE FROM categorias WHERE id = $id";
-      $mensaje = "La categoría ha sido eliminada correctamente.";
-    } else {
-      $query = "UPDATE categorias SET nombre = '$nombre' WHERE id = $id";
-      $mensaje = "La categoría ha sido actualizada correctamente.";
-    }
+  $categoria = new Categoria($conexion);
+  $categoria->setId($id);
 
+  if ($eliminar) {
+    if ($categoria->tieneHistorialAsociado()) {
+        $mensaje = "La categoría no se puede eliminar para mantener un correcto historial de préstamos, si lo desea, puede modificarla.";
+        echo "<script>alert('$mensaje');</script>";
+    } else {
+        $query = "DELETE FROM categorias WHERE id = $id";
+        $resultado = $conexion->query($query);
+
+        if ($resultado) {
+            $mensaje = "La categoría ha sido eliminada correctamente.";
+            echo "<script>alert('$mensaje'); location.href='listar.php';</script>";
+        } else {
+            // Agregar mensaje de error personalizado para la consulta de eliminación
+            $mensaje = "Error al eliminar la categoría: " . $conexion->error;
+            echo "<script>alert('$mensaje');</script>";
+        }
+    }
+  } else {
+    $query = "UPDATE categorias SET nombre = '$nombre' WHERE id = $id";
+    $mensaje = "La categoría ha sido actualizada correctamente.";
     $resultado = $conexion->query($query);
-
-    // mostrar mensaje de éxito y redirigir al listado de categorías
     echo "<script>alert('$mensaje'); location.href='listar.php';</script>";
-  } catch (mysqli_sql_exception $ex) {
-    // capturar la excepción y mostrar mensaje correspondiente
-    if (mysqli_errno($conexion) === 1451) {
-      $mensaje = "La categoría no se puede eliminar para mantener un correcto historial de préstamos, si lo desea, puede modificarla.";
-    } else {
-      $mensaje = "Ha ocurrido un error al actualizar/eliminar la categoría.";
-    }
-    echo "<script>alert('$mensaje');</script>";
   }
 }
 
@@ -71,52 +76,77 @@ if ($resultado->num_rows > 0) {
 <html>
 <head>
   <title>Editar Categorías</title>
-  <!-- Biblioteca de estilos de Bootstrap -->
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Y6XfQSTaUltcXlUhclhP" crossorigin="anonymous">
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
+  <link rel="stylesheet" href="../../assets/css/styles.css"> 
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.0/css/all.min.css" integrity="sha384-KyZXEAg3QhqLMpG8r+Knujsl5z91fgXvkR8M7r3z8GOwkpE2q3qjF3gmfsEY31pJ" crossorigin="anonymous">
 
-  <!-- Tu archivo de estilos CSS -->
-  <link rel="stylesheet" href="../assets/css/styles.css">
+
+  <style>
+    body {
+      background-image: linear-gradient(rgba(255, 255, 255, 0.5), rgba(255, 255, 255, 0.5)), url("../../assets/images/estante-librosBonita.png");
+      background-size: cover;
+      background-position: center;
+    }
+    .bg-opacity {
+      background-color: rgba(255, 255, 255, 0.8);
+      border-radius: 10px;
+      padding: 20px;
+      max-width: 100%;
+      margin: auto;
+    }
+
+    .table-container {
+      overflow-y: scroll;
+      height: 600px;
+    }
+    thead th {
+      position: sticky;
+      top: 0;
+      background-color: #fff;
+      z-index: 1;
+    }
+  </style>
 </head>
 <body>
-<div class="editar background-wrapper">
-    <div class="container editar-main-container">
-      <h2 class="titulo-centrado">Editar Categorías</h2>
-      <table>
-  <thead>
+  <div class="container-fluid">
+    <div class="row justify-content-center">
+      <div class="col-12 col-md-8 col-lg-6 p-5 bg-opacity">
+        <h2>Editar Categorías</h2>
+        <div class="table-container">
+          <table class="table table-striped">
+          <thead class="sticky-header">
+  <tr>
+    <th>ID</th>
+    <th>Nombre</th>
+    <th>Acciones</th>
+  </tr>
+</thead>
+<tbody>
+  <?php foreach ($categorias as $categoria): ?>
     <tr>
-      <th>ID</th>
-      <th>Nombre</th>
-      <th>Acciones</th>
+      <form method="POST" action="editar.php">
+        <input type="hidden" name="id" value="<?= $categoria['id'] ?>">
+        <td><?= $categoria['id'] ?></td>
+        <td><input type="text" name="nombre" class="form-control" value="<?= $categoria['nombre'] ?>"></td>
+        <td>
+          <button type="submit" name="actualizar" class="btn btn-success">Guardar</button>
+          <button type="submit" name="eliminar" class="btn btn-danger" onclick="return confirm('¿Está seguro de que desea eliminar esta categoría?')">Eliminar</button>
+        </td>
+      </form>
     </tr>
-  </thead>
-  <tbody>
-    <?php foreach ($categorias as $categoria): ?>
-      <tr>
-        <form method="POST" action="editar.php">
-          <input type="hidden" name="id" value="<?= $categoria['id'] ?>">
-          <td class="id-column"><?= $categoria['id'] ?></td>
-          <td><input type="text" name="nombre" value="<?= $categoria['nombre'] ?>"></td>
-          <td>
-            <button type="submit" name="actualizar">Guardar</button>
-            <button type="submit" name="eliminar" onclick="return confirm('¿Está seguro de que desea eliminar esta categoría?')">Eliminar</button>
-          </td>
-        </form>
-      </tr>
-    <?php endforeach ?>
-  </tbody>
+  <?php endforeach ?>
+</tbody>
 </table>
-
 </div>
-
-  </div>
-  <script>
-    <?php if (isset($_POST['actualizar'])): ?>
-      alert('La categoría ha sido actualizada correctamente.');
-      location.href = 'listar.php'; // redirigir al listado
-    <?php endif; ?>
-  </script>
+</div>
+</div>
+</div>
+<script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.min.js"></script>
 </body>
 </html>
+
+
